requirements.txt
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==1.2.0
Fiona==1.9.4.post1
fonttools==4.42.1
fpdf2==2.7.5
geopandas==0.14.0
h11==0.14.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
selenium==4.12.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webdriver-manager==4.0.0
wget==3.2
wsproto==1.2.0
